refactor: make composing easier to understand
