+++ crates/delta/src/delta_rope.rs
@@ -12,238 +12,13 @@ use self::rle_tree::Len;
 
 use super::iter::Iter;
 
+mod compose;
 pub(crate) mod rle_tree;
 
 impl<V: DeltaValue, Attr: DeltaAttr> DeltaRope<V, Attr> {
     pub fn new() -> Self {
         Self {
             tree: Default::default(),
-        }
-    }
-
-    pub fn compose(&mut self, other: &Self) {
-        // TODO: Need to implement a slow mode that is guaranteed to be correct, then we can fuzz on it
-        if self.is_empty() {
-            *self = other.clone();
-            return;
-        }
-
-        let mut index = 0;
-
-        let mut push_rest = false;
-        for item in other.iter() {
-            if index >= self.len() {
-                self.push_retain(index - self.len(), Attr::default());
-                push_rest = true;
-            }
-
-            if push_rest {
-                self.push(item.clone());
-                continue;
-            }
-
-            match item {
-                DeltaItem::Retain { len, attr } => {
-                    if self.len() < index + len {
-                        self.push_retain(index + len - self.len(), Default::default());
-                    }
-                    if !attr.attr_is_empty() {
-                        self.update_range(index..index + len, attr);
-                    }
-                    index += len;
-                }
-                DeltaItem::Replace {
-                    value: this_value,
-                    attr: this_attr,
-                    delete,
-                } => {
-                    let mut should_insert = this_value.rle_len() > 0;
-                    let mut left_del_len = *delete;
-                    if *delete > 0 {
-                        assert!(index < self.len());
-                        let range = index..(index + left_del_len).min(self.len());
-                        let from = self.tree.query::<LengthFinder>(&range.start).unwrap();
-                        let to = self.tree.query::<LengthFinder>(&range.end).unwrap();
-                        if from.cursor.leaf == to.cursor.leaf {
-                            should_insert = false;
-                            self.tree.update_leaf(from.cursor.leaf, |item| match item {
-                                DeltaItem::Retain {
-                                    len: retain_len,
-                                    attr,
-                                } => {
-                                    let start = from.cursor.offset;
-                                    let end = to.cursor.offset;
-                                    let (l, r) = match (start == 0, end >= *retain_len) {
-                                        (true, true) => {
-                                            *item = DeltaItem::Replace {
-                                                delete: left_del_len,
-                                                value: this_value.clone(),
-                                                attr: this_attr.clone(),
-                                            };
-                                            (None, None)
-                                        }
-                                        (true, false) => {
-                                            let right = item.slice(end..);
-                                            *item = DeltaItem::Replace {
-                                                delete: left_del_len,
-                                                value: this_value.clone(),
-                                                attr: this_attr.clone(),
-                                            };
-                                            (Some(right), None)
-                                        }
-                                        (false, true) => {
-                                            *retain_len -= *delete;
-                                            (
-                                                Some(DeltaItem::Replace {
-                                                    value: this_value.clone(),
-                                                    attr: this_attr.clone(),
-                                                    delete: left_del_len,
-                                                }),
-                                                None,
-                                            )
-                                        }
-                                        (false, false) => {
-                                            let right = DeltaItem::Retain {
-                                                len: *retain_len - end,
-                                                attr: attr.clone(),
-                                            };
-                                            *retain_len = start;
-                                            (
-                                                Some(DeltaItem::Replace {
-                                                    value: this_value.clone(),
-                                                    attr: this_attr.clone(),
-                                                    delete: left_del_len,
-                                                }),
-                                                Some(right),
-                                            )
-                                        }
-                                    };
-
-                                    left_del_len = 0;
-                                    (true, l, r)
-                                }
-                                DeltaItem::Replace {
-                                    value,
-                                    attr,
-                                    delete,
-                                } => {
-                                    let start = from.cursor.offset;
-                                    let end = to.cursor.offset;
-                                    let value_len = value.rle_len();
-                                    {
-                                        left_del_len =
-                                            left_del_len.saturating_sub(value_len.min(end) - start);
-                                        *delete += left_del_len;
-                                        left_del_len = 0;
-                                    }
-
-                                    let mut right = value.split(start);
-                                    right.slice_(value_len.min(end) - start..);
-                                    if this_value.rle_len() > 0 {
-                                        if attr != this_attr || !value.can_merge(this_value) {
-                                            let right = if right.rle_len() > 0 {
-                                                Some(DeltaItem::Replace {
-                                                    value: right,
-                                                    attr: attr.clone(),
-                                                    delete: 0,
-                                                })
-                                            } else {
-                                                None
-                                            };
-
-                                            return (
-                                                true,
-                                                Some(DeltaItem::Replace {
-                                                    value: this_value.clone(),
-                                                    attr: this_attr.clone(),
-                                                    delete: 0,
-                                                }),
-                                                right,
-                                            );
-                                        } else {
-                                            value.merge_right(this_value);
-                                        }
-                                    }
-
-                                    if value.can_merge(&right) {
-                                        value.merge_right(&right);
-                                        (true, None, None)
-                                    } else {
-                                        let right = if right.rle_len() > 0 {
-                                            Some(DeltaItem::Replace {
-                                                value: right,
-                                                attr: attr.clone(),
-                                                delete: 0,
-                                            })
-                                        } else {
-                                            None
-                                        };
-                                        (true, right, None)
-                                    }
-                                }
-                            });
-                        } else {
-                            self.tree.update(from.cursor..to.cursor, &mut |item| {
-                                if left_del_len == 0 {
-                                    return None;
-                                }
-
-                                match item {
-                                    DeltaItem::Retain { len, .. } => {
-                                        assert!(*len <= left_del_len);
-                                        left_del_len -= *len;
-                                        let diff = -(*len as isize);
-                                        *item = DeltaItem::Replace {
-                                            delete: *len,
-                                            value: Default::default(),
-                                            attr: Default::default(),
-                                        };
-                                        Some(Len {
-                                            data_len: diff,
-                                            delta_len: 0,
-                                        })
-                                    }
-                                    DeltaItem::Replace { value, attr, .. } => {
-                                        if left_del_len >= value.rle_len() {
-                                            let diff = value.rle_len() as isize;
-                                            left_del_len -= value.rle_len();
-                                            *value = Default::default();
-                                            *attr = Default::default();
-                                            Some(Len {
-                                                data_len: -diff,
-                                                delta_len: -diff,
-                                            })
-                                        } else {
-                                            unreachable!()
-                                        }
-                                    }
-                                }
-                            });
-                        }
-                    }
-
-                    if left_del_len > 0 || should_insert {
-                        self.insert_values(
-                            index,
-                            [DeltaItem::Replace {
-                                value: if should_insert {
-                                    this_value.clone()
-                                } else {
-                                    Default::default()
-                                },
-                                attr: if should_insert {
-                                    this_attr.clone()
-                                } else {
-                                    Default::default()
-                                },
-                                delete: left_del_len,
-                            }],
-                        );
-                    }
-
-                    index += this_value.rle_len();
-                }
-            }
         }
     }
